requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
et-xmlfile==1.1.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
gdown==5.2.0
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.65.1
h5py==3.11.0
idna==3.7
keras==2.14.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scipy==1.14.0
six==1.16.0
soupsieve==2.5
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.14.1